runner: pass bf16 logits to sample(), drop the eager pre-sample fp32 cast
Before calling sample() on the decode path, the runner eagerly upcast the
logits to fp32, dispatching one extra device program (convert) per step on
the host critical path. sample() already upcasts to fp32 internally for
the sampling transforms, and argmax over model-dtype (bf16) logits is
identical to argmax over their exact fp32 upcast, so sampled and greedy
tokens are unchanged.

The runner now forwards compute_logits' bf16 output straight into sample()
and the upcast fuses into the sample program. The sampling precompile
dummy is switched to bf16 to match the serving-time dtype (the convention
_precompile_compute_logits already uses), so serving keeps hitting the
precompiled program.

Test Plan: parametrized test_processed_logprobs_with_topk over
{float32, bfloat16} so the processed-logprobs path is exercised with the
model-dtype logits the runner now forwards.

+++ tpu_inference/runner/compilation_manager.py
@@ -892,7 +892,9 @@ class CompilationManager:
             # function.
             sampling_metadata_sharding = NamedSharding(
                 self.runner.mesh, PartitionSpec(ShardingAxisName.ATTN_DATA))
-            logits = self._create_dummy_tensor((num_reqs, hsize), jnp.float32,
+            # bf16 to match the model-dtype logits sample() receives at
+            # serving time; sample() upcasts to fp32 internally.
+            logits = self._create_dummy_tensor((num_reqs, hsize), jnp.bfloat16,
                                                logits_sharding)
             for do_sampling in (True, False):
                 for logprobs in (True, False):
